Use AcceptMessages for java.interop and add conversions between Java types and Truffle types
