+++ truffle/com.oracle.truffle.api.interop.java/src/com/oracle/truffle/api/interop/java/JavaInterop.java
@@ -24,7 +24,8 @@
  */
 package com.oracle.truffle.api.interop.java;
 
-import com.oracle.truffle.api.CallTarget;
+import java.lang.reflect.Method;
+
 import com.oracle.truffle.api.Truffle;
 import com.oracle.truffle.api.TruffleLanguage;
 import com.oracle.truffle.api.frame.VirtualFrame;
@@ -33,16 +34,7 @@ import com.oracle.truffle.api.interop.InteropException;
 import com.oracle.truffle.api.interop.Message;
 import com.oracle.truffle.api.interop.TruffleObject;
 import com.oracle.truffle.api.interop.UnsupportedMessageException;
-import com.oracle.truffle.api.nodes.Node;
 import com.oracle.truffle.api.nodes.RootNode;
-import java.lang.reflect.InvocationHandler;
-import java.lang.reflect.Method;
-import java.lang.reflect.ParameterizedType;
-import java.lang.reflect.Proxy;
-import java.lang.reflect.Type;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
 
 /**
  * Helper methods to simplify access to objects of {@link TruffleLanguage Truffle languages} from
@@ -94,7 +86,6 @@ import java.util.List;
  * <p>
  */
 public final class JavaInterop {
-    static final Object[] EMPTY = {};
 
     private JavaInterop() {
     }
@@ -142,35 +133,9 @@ public final class JavaInterop {
      *         was <code>null</code>
      */
     public static <T> T asJavaObject(Class<T> type, TruffleObject foreignObject) {
-        return asJavaObject(type, null, foreignObject);
-    }
-
-    private static <T> T asJavaObject(Class<T> clazz, Type type, TruffleObject foreignObject) {
-        Object obj;
-        if (clazz.isInstance(foreignObject)) {
-            obj = foreignObject;
-        } else {
-            if (!clazz.isInterface()) {
-                throw new IllegalArgumentException();
-            }
-            if (foreignObject == null) {
-                return null;
-            }
-            if (clazz == List.class && Boolean.TRUE.equals(binaryMessage(Message.HAS_SIZE, foreignObject))) {
-                Class<?> elementType = Object.class;
-                if (type instanceof ParameterizedType) {
-                    ParameterizedType parametrizedType = (ParameterizedType) type;
-                    final Type[] arr = parametrizedType.getActualTypeArguments();
-                    if (arr.length == 1 && arr[0] instanceof Class) {
-                        elementType = (Class<?>) arr[0];
-                    }
-                }
-                obj = TruffleList.create(elementType, foreignObject);
-            } else {
-                obj = Proxy.newProxyInstance(clazz.getClassLoader(), new Class<?>[]{clazz}, new TruffleHandler(foreignObject));
-            }
-        }
-        return clazz.cast(obj);
+        RootNode root = new TemporaryConvertRoot(TruffleLanguage.class, new ToJavaNode(), foreignObject, type);
+        Object convertedValue = Truffle.getRuntime().createCallTarget(root).call();
+        return type.cast(convertedValue);
     }
 
     /**
@@ -228,12 +193,8 @@ public final class JavaInterop {
      * @return instance of interface that wraps the provided <code>function</code>
      */
     public static <T> T asJavaFunction(Class<T> functionalType, TruffleObject function) {
-        final Method[] arr = functionalType.getDeclaredMethods();
-        if (!functionalType.isInterface() || arr.length != 1) {
-            throw new IllegalArgumentException();
-        }
-        Object obj = Proxy.newProxyInstance(functionalType.getClassLoader(), new Class<?>[]{functionalType}, new SingleHandler(function));
-        return functionalType.cast(obj);
+        RootNode root = new TemporaryConvertRoot(TruffleLanguage.class, new ToJavaNode(), function, functionalType);
+        return functionalType.cast(Truffle.getRuntime().createCallTarget(root).call());
     }
 
     /**
@@ -264,301 +225,22 @@ public final class JavaInterop {
         return new JavaFunctionObject(arr[0], implementation);
     }
 
-    static Message findMessage(MethodMessage mm) {
-        if (mm == null) {
-            return null;
-        }
-        return Message.valueOf(mm.message());
-    }
-
-    static Object toJava(Object ret, Method method) {
-        Class<?> retType = method.getReturnType();
-        Object primitiveRet = toPrimitive(ret, retType);
-        if (primitiveRet != null) {
-            return primitiveRet;
-        }
-        if (ret instanceof TruffleObject) {
-            if (Boolean.TRUE.equals(binaryMessage(Message.IS_NULL, ret))) {
-                return null;
-            }
-        }
-        if (retType.isInstance(ret)) {
-            return ret;
-        }
-        if (ret instanceof TruffleObject) {
-            final TruffleObject truffleObject = (TruffleObject) ret;
-            if (retType.isInterface()) {
-                return asJavaObject(retType, method.getGenericReturnType(), truffleObject);
-            }
-        }
-        return ret;
-    }
-
-    private static final class SingleHandler implements InvocationHandler {
-        private final TruffleObject symbol;
-        private CallTarget target;
-
-        SingleHandler(TruffleObject obj) {
-            this.symbol = obj;
-        }
-
-        @Override
-        public Object invoke(Object proxy, Method method, Object[] arguments) throws Throwable {
-            Object[] args = arguments == null ? EMPTY : arguments;
-            if (target == null) {
-                Node executeMain = Message.createExecute(args.length).createNode();
-                RootNode symbolNode = new TemporaryRoot(TruffleLanguage.class, executeMain, symbol);
-                target = Truffle.getRuntime().createCallTarget(symbolNode);
-            }
-            Object ret = target.call(args);
-            return toJava(ret, method);
-        }
-    }
-
-    private static final class TruffleHandler implements InvocationHandler {
-        private final TruffleObject obj;
-
-        TruffleHandler(TruffleObject obj) {
-            this.obj = obj;
-        }
-
-        @Override
-        public Object invoke(Object proxy, Method method, Object[] arguments) throws Throwable {
-            Object[] args = arguments == null ? EMPTY : arguments;
-            Object val;
-            for (int i = 0; i < args.length; i++) {
-                if (args[i] == null) {
-                    continue;
-                }
-                if (Proxy.isProxyClass(args[i].getClass())) {
-                    InvocationHandler h = Proxy.getInvocationHandler(args[i]);
-                    if (h instanceof TruffleHandler) {
-                        args[i] = ((TruffleHandler) h).obj;
-                    }
-                }
-            }
-
-            if (Object.class == method.getDeclaringClass()) {
-                return method.invoke(obj, args);
-            }
-
-            String name = method.getName();
-            Message message = findMessage(method.getAnnotation(MethodMessage.class));
-            if (message == Message.WRITE) {
-                if (args.length != 1) {
-                    throw new IllegalStateException("Method needs to have a single argument to handle WRITE message " + method);
-                }
-                message(Message.WRITE, obj, name, args[0]);
-                return null;
-            }
-            if (message == Message.HAS_SIZE || message == Message.IS_BOXED || message == Message.IS_EXECUTABLE || message == Message.IS_NULL || message == Message.GET_SIZE) {
-                return message(message, obj);
-            }
-
-            if (message == Message.READ) {
-                val = message(Message.READ, obj, name);
-                return toJava(val, method);
-            }
-
-            if (message == Message.UNBOX) {
-                val = message(Message.UNBOX, obj);
-                return toJava(val, method);
-            }
-
-            if (Message.createExecute(0).equals(message)) {
-                List<Object> copy = new ArrayList<>(args.length);
-                copy.addAll(Arrays.asList(args));
-                message = Message.createExecute(copy.size());
-                val = message(message, obj, copy.toArray());
-                return toJava(val, method);
-            }
-
-            if (Message.createInvoke(0).equals(message)) {
-                List<Object> copy = new ArrayList<>(args.length + 1);
-                copy.add(name);
-                copy.addAll(Arrays.asList(args));
-                message = Message.createInvoke(args.length);
-                val = message(message, obj, copy.toArray());
-                return toJava(val, method);
-            }
-
-            if (Message.createNew(0).equals(message)) {
-                message = Message.createNew(args.length);
-                val = message(message, obj, args);
-                return toJava(val, method);
-            }
-
-            if (message == null) {
-                Object ret;
-                try {
-                    List<Object> callArgs = new ArrayList<>(args.length);
-                    callArgs.add(name);
-                    callArgs.addAll(Arrays.asList(args));
-                    ret = message(Message.createInvoke(args.length), obj, callArgs.toArray());
-                } catch (InteropException ex) {
-                    val = message(Message.READ, obj, name);
-                    Object primitiveVal = toPrimitive(val, method.getReturnType());
-                    if (primitiveVal != null) {
-                        return primitiveVal;
-                    }
-                    TruffleObject attr = (TruffleObject) val;
-                    if (Boolean.FALSE.equals(message(Message.IS_EXECUTABLE, attr))) {
-                        if (args.length == 0) {
-                            return toJava(attr, method);
-                        }
-                        throw new IllegalArgumentException(attr + " cannot be invoked with " + args.length + " parameters");
-                    }
-                    List<Object> callArgs = new ArrayList<>(args.length);
-                    callArgs.addAll(Arrays.asList(args));
-                    ret = message(Message.createExecute(callArgs.size()), attr, callArgs.toArray());
-                }
-                return toJava(ret, method);
-            }
-            throw new IllegalArgumentException("Unknown message: " + message);
-        }
-
-    }
-
-    static boolean isPrimitive(Object attr) {
-        return toPrimitive(attr, null) != null;
-    }
-
-    static Object toPrimitive(Object value, Class<?> requestedType) {
-        Object attr;
-        if (value instanceof TruffleObject) {
-            if (!Boolean.TRUE.equals(binaryMessage(Message.IS_BOXED, value))) {
-                return null;
-            }
-            try {
-                attr = message(Message.UNBOX, value);
-            } catch (InteropException e) {
-                throw new IllegalStateException();
-            }
-        } else {
-            attr = value;
-        }
-        if (attr instanceof Number) {
-            if (requestedType == null) {
-                return attr;
-            }
-            Number n = (Number) attr;
-            if (requestedType == byte.class || requestedType == Byte.class) {
-                return n.byteValue();
-            }
-            if (requestedType == short.class || requestedType == Short.class) {
-                return n.shortValue();
-            }
-            if (requestedType == int.class || requestedType == Integer.class) {
-                return n.intValue();
-            }
-            if (requestedType == long.class || requestedType == Long.class) {
-                return n.longValue();
-            }
-            if (requestedType == float.class || requestedType == Float.class) {
-                return n.floatValue();
-            }
-            if (requestedType == double.class || requestedType == Double.class) {
-                return n.doubleValue();
-            }
-            if (requestedType == char.class || requestedType == Character.class) {
-                return (char) n.intValue();
-            }
-            return n;
-        }
-        if (attr instanceof String) {
-            if (requestedType == char.class || requestedType == Character.class) {
-                if (((String) attr).length() == 1) {
-                    return ((String) attr).charAt(0);
-                }
-            }
-            return attr;
-        }
-        if (attr instanceof Character) {
-            return attr;
-        }
-        if (attr instanceof Boolean) {
-            return attr;
-        }
-        return null;
-    }
-
-    @SuppressWarnings("unused")
-    static Object message(final Message m, Object receiver, Object... arr) throws InteropException {
-        Node n = m.createNode();
-        CallTarget callTarget = Truffle.getRuntime().createCallTarget(new TemporaryRoot(TruffleLanguage.class, n, (TruffleObject) receiver));
-        return callTarget.call(arr);
-    }
-
-    static Object binaryMessage(final Message m, Object receiver, Object... arr) {
-        try {
-            return message(m, receiver, arr);
-        } catch (InteropException e) {
-            throw new AssertionError(e);
-        }
-    }
-
-    private static class TemporaryRoot extends RootNode {
-        @Node.Child private Node foreignAccess;
-        private final TruffleObject function;
+    private static class TemporaryConvertRoot extends RootNode {
+        @Child private ToJavaNode node;
+        private final Object value;
+        private final Class<?> type;
 
         @SuppressWarnings("rawtypes")
-        TemporaryRoot(Class<? extends TruffleLanguage> lang, Node foreignAccess, TruffleObject function) {
+        TemporaryConvertRoot(Class<? extends TruffleLanguage> lang, ToJavaNode node, Object value, Class<?> type) {
             super(lang, null, null);
-            this.foreignAccess = foreignAccess;
-            this.function = function;
+            this.node = node;
+            this.value = value;
+            this.type = type;
         }
 
-        @SuppressWarnings("deprecation")
         @Override
         public Object execute(VirtualFrame frame) {
-            return ForeignAccess.execute(foreignAccess, frame, function, frame.getArguments());
+            return node.convert(frame, value, type);
         }
-    } // end of TemporaryRoot
-
-    static final class JavaFunctionObject implements TruffleObject {
-        final Method method;
-        final Object obj;
-
-        JavaFunctionObject(Method method, Object obj) {
-            this.method = method;
-            this.obj = obj;
-        }
-
-        @Override
-        public ForeignAccess getForeignAccess() {
-            return ForeignAccess.create(JavaFunctionObject.class, new JavaFunctionForeignAccess());
-        }
-
-    } // end of JavaFunctionObject
-
-    static final class JavaObject implements TruffleObject {
-        static final JavaObject NULL = new JavaObject(null, Object.class);
-
-        final Object obj;
-        final Class<?> clazz;
-
-        JavaObject(Object obj, Class<?> clazz) {
-            this.obj = obj;
-            this.clazz = clazz;
-        }
-
-        @Override
-        public ForeignAccess getForeignAccess() {
-            return ForeignAccess.create(JavaObject.class, new JavaObjectForeignAccess());
-        }
-
-        @Override
-        public int hashCode() {
-            return System.identityHashCode(obj);
-        }
-
-        @Override
-        public boolean equals(Object other) {
-            if (other instanceof JavaObject) {
-                return obj == ((JavaObject) other).obj && clazz == ((JavaObject) other).clazz;
-            }
-            return false;
-        }
-    } // end of JavaObject
-
+    }
 }
