+++ truffle/com.oracle.truffle.api.interop.java/src/com/oracle/truffle/api/interop/java/JavaInteropLanguage.java
@@ -26,5 +26,5 @@ package com.oracle.truffle.api.interop.java;
 
 import com.oracle.truffle.api.TruffleLanguage;
 
-abstract class JavaInteropLanguage extends TruffleLanguage<Object> {
+public abstract class JavaInteropLanguage extends TruffleLanguage<Object> {
 }
