+++ truffle/com.oracle.truffle.api.interop.java/src/com/oracle/truffle/api/interop/java/ReadFieldNode.java
@@ -24,52 +24,57 @@
  */
 package com.oracle.truffle.api.interop.java;
 
-import com.oracle.truffle.api.frame.VirtualFrame;
-import com.oracle.truffle.api.interop.ForeignAccess;
-import com.oracle.truffle.api.nodes.RootNode;
 import java.lang.reflect.Array;
 import java.lang.reflect.Field;
 import java.lang.reflect.Method;
 import java.lang.reflect.Modifier;
 
-final class ReadFieldNode extends RootNode {
-    ReadFieldNode() {
-        super(JavaInteropLanguage.class, null, null);
+import com.oracle.truffle.api.dsl.Specialization;
+import com.oracle.truffle.api.frame.VirtualFrame;
+import com.oracle.truffle.api.interop.AcceptMessage;
+
+/**
+ * {@link ReadFieldBaseNode} is a generated class that uses {@link Specialization}.
+ */
+@AcceptMessage(value = "READ", receiverType = JavaObject.class, language = JavaInteropLanguage.class)
+final class ReadFieldNode extends ReadFieldBaseNode {
+    @Override
+    public Object access(VirtualFrame frame, JavaObject object, int index) {
+        Object obj = object.obj;
+        Object val = Array.get(obj, index);
+
+        if (ToJavaNode.isPrimitive(val)) {
+            return val;
+        }
+        return JavaInterop.asTruffleObject(val);
     }
 
     @Override
-    public Object execute(VirtualFrame frame) {
+    public Object access(VirtualFrame frame, JavaObject object, String name) {
         try {
-            JavaInterop.JavaObject receiver = (JavaInterop.JavaObject) ForeignAccess.getReceiver(frame);
-            Object obj = receiver.obj;
+            Object obj = object.obj;
             final boolean onlyStatic = obj == null;
-            final Object nameOrIndex = ForeignAccess.getArguments(frame).get(0);
             Object val;
-            if (nameOrIndex instanceof Integer) {
-                val = Array.get(obj, (int) nameOrIndex);
-            } else {
-                String name = (String) nameOrIndex;
-                try {
-                    final Field field = receiver.clazz.getField(name);
-                    final boolean isStatic = (field.getModifiers() & Modifier.STATIC) != 0;
-                    if (onlyStatic != isStatic) {
-                        throw new NoSuchFieldException();
-                    }
-                    val = field.get(obj);
-                } catch (NoSuchFieldException ex) {
-                    for (Method m : receiver.clazz.getMethods()) {
-                        final boolean isStatic = (m.getModifiers() & Modifier.STATIC) != 0;
-                        if (onlyStatic != isStatic) {
-                            continue;
-                        }
-                        if (m.getName().equals(name)) {
-                            return new JavaInterop.JavaFunctionObject(m, obj);
-                        }
-                    }
-                    throw (NoSuchFieldError) new NoSuchFieldError(ex.getMessage()).initCause(ex);
+            try {
+                final Field field = object.clazz.getField(name);
+                final boolean isStatic = (field.getModifiers() & Modifier.STATIC) != 0;
+                if (onlyStatic != isStatic) {
+                    throw new NoSuchFieldException();
                 }
+                val = field.get(obj);
+            } catch (NoSuchFieldException ex) {
+                for (Method m : object.clazz.getMethods()) {
+                    final boolean isStatic = (m.getModifiers() & Modifier.STATIC) != 0;
+                    if (onlyStatic != isStatic) {
+                        continue;
+                    }
+                    if (m.getName().equals(name)) {
+                        return new JavaFunctionObject(m, obj);
+                    }
+                }
+                throw (NoSuchFieldError) new NoSuchFieldError(ex.getMessage()).initCause(ex);
             }
-            if (JavaInterop.isPrimitive(val)) {
+            if (ToJavaNode.isPrimitive(val)) {
                 return val;
             }
             return JavaInterop.asTruffleObject(val);
